refactor: restore exported namespace merge
