--- src/sfCommandNamespace.ts
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2023, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-import { StructuredMessage } from '@salesforce/core';
-
-export namespace SfCommand {
-  export type Info = StructuredMessage | string;
-  export type Warning = StructuredMessage | string;
-
-  export interface Json<T> {
-    status: number;
-    result: T;
-    warnings?: Warning[];
-  }
-
-  export interface Error {
-    status: number;
-    name: string;
-    message: string;
-    stack: string | undefined;
-    warnings?: Warning[];
-    actions?: string[];
-    code?: unknown;
-    exitCode?: number;
-    data?: unknown;
-    context?: string;
-    commandName?: string;
-  }
-}
